move eraseAccidentalFunction, addAccidentalToNote, changeNoteFunction to one modifyNotes file
